Move js/vuex-queue.js and js/distribution -> utils/ | Move js/pubsub.js -> controller/
