+++ frontend/simple/utils/distribution/mincome-default.js
