+++ frontend/simple/utils/distribution/mincome-proportional.js
